+++ test/compliance/literals.test.js
@@ -1,3 +1,5 @@
+require('../cds')
+
 describe('literals', () => {
   describe('globals', () => {
     describe('boolean', () => {